tg_bot/requirements.txt
langchain==0.3.1
langchain-core==0.3.6
aiogram==3.13.1
aiohttp==3.9.5
python-dotenv  
pydub==0.25.1  
torch==2.5.0
transformers==4.45.1  
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
faiss-cpu==1.9.0
aiohttp==3.9.5 
scipy==1.13.0  
numpy==1.26.3  
langchain-huggingface==0.1.0
ffmpeg==1.4
ffmpeg-downloader==0.3.0
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langchain-community==0.3.1
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
accelerate>=0.26.0